- replaced int with double in all equations

+++ src/ghozti/calc/equations/BasicEquations.java
@@ -2,15 +2,15 @@ package ghozti.calc.equations;
 
 public class BasicEquations {
 
-    public static int add(int num1, int num2){
+    public static double add(double num1, double num2){
         return num1 + num2;
     }
 
-    public static int subtract(int num1, int num2){
+    public static double subtract(double num1, double num2){
         return num1 - num2;
     }
 
-    public static int divide(int num1, int num2){
+    public static double divide(double num1, double num2){
         return num1 / num2;
     }
 }
